models/2-Gleb/train/requirements_def.txt
absl-py==0.11.0
affine==2.3.0
aiocontextvars==0.2.2
albumentations==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.2.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2018.1.18
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
contextvars==2.4
cryptography==3.4.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
distro==1.5.0
docker==4.4.1
docker-compose==1.28.2
dockerpty==0.4.1
docopt==0.6.2
easydict==1.7
efficientnet-pytorch==0.6.3
entrypoints==0.3
fastprogress==1.0.0
fire==0.3.1
Flask==1.1.2
future==0.18.2
GDAL==2.2.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
idna==2.6
imageio==2.9.0
imgaug==0.4.0
immutables==0.14
importlib-metadata==3.1.0
ipykernel==5.3.4
ipympl==0.6.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
json-tricks==3.15.5
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
jupytext==1.7.1
kaggle==1.5.10
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.3
markdown-it-py==0.5.6
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nbstripout==0.3.9
nest-asyncio==1.4.3
networkx==2.5
ninja==1.10.0.post2
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
packaging==20.7
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pretrainedmodels==0.7.4
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.2
pygobject==3.26.1
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.5
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
ranger-fm==1.9.3
rasterio==1.2.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.12.0
snuggs==1.4.7
ssh-import-id==5.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.3
tifffile==2020.9.3
timm==0.3.2
toml==0.10.2
torch==1.7.0
torchsummary==1.5.1
torchvision==0.8.1
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
ttach==0.0.3
typing-extensions==3.7.4.3
urllib3==1.22
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.4.0
